Making a hit counter in PHP, the Extensive Way
How to make a hit counter. Depending on if the IP_address doesn't exist in the database already. There are functions that add the ip_address to the database, and it increments the counter once. Then it checks to see if the same ip_address enters the page if so, it doesn't increment the counter.